Add AdvantageKit logging
Added AdvantageKit for extra diagnostics and logging for the robot. This will help hook into the AdvantageScope tool for WPILib.

Added a Git metadata information task for each Gradle build, such as the commit SHA, the date of the build, the branch, and if the build was dirty (has changes not committed). This requires the installation of the Git CLI and be available on the PATH.

Fixed issue where formatting failed a build. This will now perform the formatting on all builds, including deploy jobs.