Fix crash if replay log not found, add robot software metadata
Fixed an issue where if not provided a log of a match during robot simulation, the software would crash. This now is handled with just running the simulation as normal without a replay. There is no current logic for how to feed information into the simulator at the moment.

Added metadata for robot software. This includes runtime type such as a simulation or a real robot, as well as Git information such as the commit SHA, the build date, the commit date, branch, and whether or not if there are files not committed to the repository.