Fix logfile null crash, add Git warning to driver station
Fixed an issue that where if an AdvantageKit logfile was not provided or was null during the robot simulation startup there would be a simulator crash. This is now fixed.

Added Git warnings to the driver station. This will notify the driver if there are currently uncommitted changes on the robot as well as if the robot is currently not on the "main" branch. This will show up as errors as we are treating these warnings as errors.