Reformatted Input Detections
Reformatted to make the if (true)'s less lengthy and messy. Additionally, the fly wheels now reverse from button 11 AND 7 (pulled it closer). One more thing, the conveyor can now be controlled separately from other parts of the intake by using button 9 to reverse it and button 10 to advance it.